requirements.txt
Django==1.10.3
django-randomslugfield==0.3.0
django-tinymce==2.4.0
MySQL-python==1.2.5
